+++ src/main/java/com/example/demo/service/ProductService.java
@@ -23,11 +23,11 @@ public class ProductService {
         productRepository.save(product);
     }
 
-    public Product getProductById(int id) {
+    public Product getProductById(Long id) {
         return productRepository.findById(id).orElse(null);
     }
 
-    public void deleteProduct(int id) {
+    public void deleteProduct(Long id) {
         productRepository.deleteById(id);
     }
 }
